Update dependencies and refactor media metadata handling. Added markdown support and simplified generic_metadata function for improved clarity and efficiency.
