+++ crates/utils/dependent/Cargo.toml
@@ -20,6 +20,7 @@ file-storage-service = { path = "../../services/file-storage" }
 fitness-models = { path = "../../models/fitness" }
 importer-models = { path = "../../models/importer" }
 itertools = { workspace = true }
+markdown = { workspace = true }
 media-models = { path = "../../models/media" }
 migrations = { path = "../../migrations" }
 nanoid = { workspace = true }
